Added Logic for the indexer, arm, and shooter
Added command for the intake button.
When the arm is down, the button makes the intake intake and the indexer rotate until the beam break is broken.
When the arm is in shooter mode, the button makes the indexer feed the shooter.
When the arm is in the amp mode, the button makes the indexer outtake into the amp.